+++ scripts/constants.js
@@ -5,6 +5,8 @@ const CONSTANTS = {
         ATTACHED_TEMPLATE: 'attachedTemplate',
         ATTACHED_REGION: 'attachedRegion',
         ATTACH_REGION_TO_TEMPLATE: 'attachRegionToTemplate',
+        IS_CONFIG_REGION: 'isConfigRegion',
+        REGION_BEHAVIORS: 'regionBehaviors',
     },
 };