+++ changelog.md
@@ -1,5 +1,9 @@
 # Region Attacher Changelog
 
+## Version 0.4.0
+- Region Configuration Menu added; now if attaching a Scene Region to an item-created template, can pre-configure its behaviors in the item sheet
+    - For now, this is restricted only to GMs, as a temporary region is created in the background while the behaviors are pre-configured
+
 ## Version 0.3.1
 - Added tidy5e integration